requirements.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.13.1
Pygments==2.11.2
pytz==2021.3
sqlparse==0.4.2
